Merged my 'advice' to upstream.
I use sql-indent on my work, and sometimes sql-indent does not work as
I expected, for example:

1. Indentation does not work for some keywords by default:
   BEGIN, ELSE, END, DELIMITER and so on.

2. "TAB" behaves differently compared to other modes when cursor in
placed at beginning of line:
it indents code but does not move cursor accordingly. Suppose I move
cursor to the beginning of next line which is empty, and then I want
to stroke TAB to move cursor to correct indentation before typing SQL
statements, it does nothing. I have to type something, and then stroke
TAB to make it work. This is different compared to all other modes.

I use several `advice' to make sql-indent work as I expected, but it
would be great if this could be done internally in sql-indent package.

Thanks.